WP 5.6 upgrade and state autocomplete fix

+++ public_html/wp-content/plugins/nextend-facebook-connect/NSL/Persistent/Persistent.php
@@ -42,17 +42,23 @@ class Persistent {
     }
 
     public static function set($key, $value) {
-
-        self::$instance->storage->set($key, $value);
+        if (self::$instance->storage) {
+            self::$instance->storage->set($key, $value);
+        }
     }
 
     public static function get($key) {
+        if (self::$instance->storage) {
+            return self::$instance->storage->get($key);
+        }
 
-        return self::$instance->storage->get($key);
+        return false;
     }
 
     public static function delete($key) {
-        self::$instance->storage->delete($key);
+        if (self::$instance->storage) {
+            self::$instance->storage->delete($key);
+        }
     }
 
     /**
@@ -77,8 +83,9 @@ class Persistent {
     }
 
     public static function clear() {
-
-        self::$instance->storage->clear();
+        if (self::$instance->storage) {
+            self::$instance->storage->clear();
+        }
     }
 }
